requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
aif360==0.5.0
tensorflow==2.15.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
joblib==1.3.2
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.13.0
dvc==3.40.0
mlflow==2.9.2
dagshub==0.3.12